quotient: move insertion assertion for unique items
Previously, if the quotient filter is full and an item was inserted into
the filter, an assertion would fire even if the item already existed in
the filter. This commit moves the assertion to fire after the duplicate
check.

This commit also tweaks the stress test for quotient filter to insert a
deterministic amount of items into the filter.

Fixes #2.