Fix insufficient randomness in CuckooFilter::insert
The CuckooFilter generates random numbers in the insert operation.
However, because it creates a new PRNG from the same seed each time, it
will always choose the same initial index. Furthermore, when inserting
many entries with the same hash, it will probe each index in the same
order, resulting in poor performance.

Instead of reconstructing the PRNG on each insert, only construct it
once for the lifetime of the CuckooFilter.

Fixes #3
https://gitlab.com/jeffrey-xiao/probabilistic-collections-rs/-/issues/3