cuckoo: optimize fingerprint storage
A terrible design decision was made for `BitArrayVec` to return
`Vec<u8>`. The initial goal of `BitArrayVec` was to store bit arrays of
arbitrary length. For the purposes of this crate, we were only concerned
about bit arrays with length less than equal to 64. This commit fixes
this problem and does the following:

- Replace `BitArrayVec` with `BitstringVec` which returns `u64`
  representing a bitstring of length at most 64.
- Remove `BitVec` and `BitstringVec` from the public API since they were
  meant to only be used internally.
- Refactor `CuckooFilter` to use `BitstringVec`.
- Refactor `QuotientFilter` to use `BitstringVec`.